+++ src/main/java/org/findy/findy_be/auth/api/AuthController.java
@@ -1,9 +1,9 @@
-package org.findy.findy_be.auth.controller;
+package org.findy.findy_be.auth.api;
 
 import java.util.Date;
 
-import org.findy.findy_be.auth.entity.AuthRequestModel;
-import org.findy.findy_be.auth.oauth.entity.UserPrincipal;
+import org.findy.findy_be.auth.dto.AuthRequestModel;
+import org.findy.findy_be.auth.oauth.domain.UserPrincipal;
 import org.findy.findy_be.auth.oauth.token.AuthToken;
 import org.findy.findy_be.auth.oauth.token.AuthTokenProvider;
 import org.findy.findy_be.common.config.AppProperties;
