+++ src/main/java/org/findy/findy_be/auth/oauth/service/CustomOAuth2UserService.java
@@ -2,8 +2,8 @@ package org.findy.findy_be.auth.oauth.service;
 
 import java.time.LocalDateTime;
 
-import org.findy.findy_be.auth.oauth.entity.SocialProviderType;
-import org.findy.findy_be.auth.oauth.entity.UserPrincipal;
+import org.findy.findy_be.auth.oauth.domain.SocialProviderType;
+import org.findy.findy_be.auth.oauth.domain.UserPrincipal;
 import org.findy.findy_be.auth.oauth.info.OAuth2UserInfo;
 import org.findy.findy_be.auth.oauth.info.OAuth2UserInfoFactory;
 import org.findy.findy_be.common.exception.custom.OAuthProviderMissMatchException;
