+++ src/main/java/org/findy/findy_be/auth/oauth/service/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package org.findy.findy_be.auth.oauth.service;
 
 import static org.findy.findy_be.common.exception.ErrorCode.*;
 
-import org.findy.findy_be.auth.oauth.entity.UserPrincipal;
+import org.findy.findy_be.auth.oauth.domain.UserPrincipal;
 import org.findy.findy_be.user.entity.User;
 import org.findy.findy_be.user.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
