+++ src/main/java/org/findy/findy_be/user/api/UserController.java
@@ -1,8 +1,8 @@
-package org.findy.findy_be.user.controller;
+package org.findy.findy_be.user.api;
 
 import org.findy.findy_be.common.meta.LoginUser;
+import org.findy.findy_be.user.application.UserService;
 import org.findy.findy_be.user.entity.User;
-import org.findy.findy_be.user.service.UserService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
